+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- `counter`: Added overflow latch
+
 ## 1.13.1 - 2019-06-01
 
 ### Changed